Refactor SwiftPromises into a frameworrk, and renaming the app to SwiftPromises Demo
