+++ SwiftPromises Demo/AppDelegate.swift
