+++ SwiftPromises Demo/Base.lproj/Main.storyboard
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="6254" systemVersion="14C109" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" initialViewController="rJ4-8q-dyD">
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="7528.3" systemVersion="14C1514" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" initialViewController="rJ4-8q-dyD">
     <dependencies>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="6247"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="7518.3"/>
     </dependencies>
     <scenes>
         <!--Navigation Controller-->
@@ -118,7 +118,7 @@
         <!--Simple load from network-->
         <scene sceneID="bt5-Ub-hSE">
             <objects>
-                <viewController id="BP2-5N-jER" customClass="SimpleNetworkCallDemoViewController" customModule="SwiftPromises" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController id="BP2-5N-jER" customClass="SimpleNetworkCallDemoViewController" customModule="SwiftPromises_Demo" customModuleProvider="target" sceneMemberID="viewController">
                     <layoutGuides>
                         <viewControllerLayoutGuide type="top" id="JUh-lN-gnS"/>
                         <viewControllerLayoutGuide type="bottom" id="jza-sv-zb0"/>
@@ -205,7 +205,7 @@
         <!--Timeout network call-->
         <scene sceneID="gxQ-UQ-kdO">
             <objects>
-                <viewController id="dl3-d2-Pma" customClass="TimeoutNetworkCallDemoViewController" customModule="SwiftPromises" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController id="dl3-d2-Pma" customClass="TimeoutNetworkCallDemoViewController" customModule="SwiftPromises_Demo" customModuleProvider="target" sceneMemberID="viewController">
                     <layoutGuides>
                         <viewControllerLayoutGuide type="top" id="KSC-17-vNT"/>
                         <viewControllerLayoutGuide type="bottom" id="6CZ-iT-0A1"/>
@@ -290,7 +290,7 @@
         <!--Chained load data from network-->
         <scene sceneID="d2E-k7-GM8">
             <objects>
-                <viewController id="Oh8-Vz-zUd" userLabel="Chained load data from network" customClass="ChainedNetworkCallDemoViewController" customModule="SwiftPromises" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController id="Oh8-Vz-zUd" userLabel="Chained load data from network" customClass="ChainedNetworkCallDemoViewController" customModule="SwiftPromises_Demo" customModuleProvider="target" sceneMemberID="viewController">
                     <layoutGuides>
                         <viewControllerLayoutGuide type="top" id="YRB-qJ-0fK"/>
                         <viewControllerLayoutGuide type="bottom" id="AJ8-0Q-mJB"/>
@@ -403,7 +403,7 @@
         <!--Compound load data from network-->
         <scene sceneID="5IK-eC-W9s">
             <objects>
-                <viewController id="l73-Fj-TVr" userLabel="Compound load data from network" customClass="CompoundNetworkCallDemoViewController" customModule="SwiftPromises" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController id="l73-Fj-TVr" userLabel="Compound load data from network" customClass="CompoundNetworkCallDemoViewController" customModule="SwiftPromises_Demo" customModuleProvider="target" sceneMemberID="viewController">
                     <layoutGuides>
                         <viewControllerLayoutGuide type="top" id="dhq-Qn-CA5"/>
                         <viewControllerLayoutGuide type="bottom" id="aJ0-3y-mSC"/>
