+++ SwiftPromises Demo/ChainedNetworkCallDemoViewController.swift
@@ -1,5 +1,5 @@
 //
-//  CompoundNetworkCallDemoViewController.swift
+//  ChainedNetworkCallDemoViewController.swift
 //  SwiftPromises
 //
 //  Created by Douglas Sjoquist on 3/1/15.
@@ -7,28 +7,22 @@
 //
 
 import UIKit
+import SwiftPromises
 
-class CompoundNetworkCallDemoViewController: BaseDemoViewController {
+class ChainedNetworkCallDemoViewController: BaseDemoViewController {
 
     @IBOutlet var url1TextField:UITextField?
-    @IBOutlet var url2aTextField:UITextField?
-    @IBOutlet var url2bTextField:UITextField?
+    @IBOutlet var url2TextField:UITextField?
     @IBOutlet var url3TextField:UITextField?
-    @IBOutlet var delay2aStepper:UIStepper?
-    @IBOutlet var delay2bStepper:UIStepper?
-    @IBOutlet var delay2aLabel:UILabel?
-    @IBOutlet var delay2bLabel:UILabel?
     @IBOutlet var url1StatusImageView:UIImageView?
-    @IBOutlet var url2aStatusImageView:UIImageView?
-    @IBOutlet var url2bStatusImageView:UIImageView?
+    @IBOutlet var url2StatusImageView:UIImageView?
     @IBOutlet var url3StatusImageView:UIImageView?
     @IBOutlet var finalStatusImageView:UIImageView?
     @IBOutlet var stopOnErrorSwitch:UISwitch?
 
     override func viewWillAppear(animated: Bool) {
         url1TextField!.text = "http://cnn.com"
-        url2aTextField!.text = "http://apple.com"
-        url2bTextField!.text = "http://dilbert.com"
+        url2TextField!.text = "http://apple.com"
         url3TextField!.text = "http://nytimes.com"
 
         super.viewWillAppear(animated)
@@ -36,8 +30,7 @@ class CompoundNetworkCallDemoViewController: BaseDemoViewController {
 
     override func clearStatus() {
         url1StatusImageView!.setStatus(nil)
-        url2aStatusImageView!.setStatus(nil)
-        url2bStatusImageView!.setStatus(nil)
+        url2StatusImageView!.setStatus(nil)
         url3StatusImageView!.setStatus(nil)
         finalStatusImageView!.setStatus(nil)
     }
@@ -46,13 +39,6 @@ class CompoundNetworkCallDemoViewController: BaseDemoViewController {
         return true
     }
 
-    override func updateUI() {
-        super.updateUI()
-
-        delay2aLabel!.text =  "Delay2a: \(delay2aStepper!.value)"
-        delay2bLabel!.text =  "Delay2b: \(delay2bStepper!.value)"
-    }
-    
     override func start() {
         clearStatus()
         clearLog()
@@ -61,13 +47,7 @@ class CompoundNetworkCallDemoViewController: BaseDemoViewController {
 
         loadURL1StepPromise().then(
             { [weak self] (value) -> AnyObject? in
-                self?.log("loading promise2a and promise2b")
-                var promises:[Promise] = []
-                if let strongSelf = self {
-                    promises.append(strongSelf.loadURL2aStepPromise())
-                    promises.append(strongSelf.loadURL2bStepPromise())
-                }
-                return Promise.all(promises)
+                return self?.loadURL2StepPromise()
         } ).then(
             { [weak self] (value) -> AnyObject? in
                 return self?.loadURL3StepPromise()
@@ -87,24 +67,20 @@ class CompoundNetworkCallDemoViewController: BaseDemoViewController {
     }
 
     func loadURL1StepPromise() -> Promise {
-        return loadURLStepPromise(url1TextField!.text, statusImageView:url1StatusImageView!, delay:0.0)
+        return loadURLStepPromise(url1TextField!.text, statusImageView:url1StatusImageView!)
     }
 
-    func loadURL2aStepPromise() -> Promise {
-        return loadURLStepPromise(url2aTextField!.text, statusImageView:url2aStatusImageView!, delay:delay2aStepper!.value)
-    }
-
-    func loadURL2bStepPromise() -> Promise {
-        return loadURLStepPromise(url2bTextField!.text, statusImageView:url2bStatusImageView!, delay:delay2bStepper!.value)
+    func loadURL2StepPromise() -> Promise {
+        return loadURLStepPromise(url2TextField!.text, statusImageView:url2StatusImageView!)
     }
 
     func loadURL3StepPromise() -> Promise {
-        return loadURLStepPromise(url3TextField!.text, statusImageView:url3StatusImageView!, delay:0.0)
+        return loadURLStepPromise(url3TextField!.text, statusImageView:url3StatusImageView!)
     }
 
-    func loadURLStepPromise(urlString:String?, statusImageView:UIImageView?, delay:NSTimeInterval) -> Promise {
+    func loadURLStepPromise(urlString:String?, statusImageView:UIImageView?) -> Promise {
         var url:NSURL? = (urlString == nil) ? nil : NSURL(string:urlString!)
-        return loadURLPromise(url, delay:delay).then(
+        return loadURLPromise(url).then(
             { [weak self] (value) -> AnyObject? in
                 statusImageView?.setStatus(true)
                 var data:NSData? = value as? NSData
@@ -113,7 +89,7 @@ class CompoundNetworkCallDemoViewController: BaseDemoViewController {
             }, reject: { [weak self] (error) -> AnyObject? in
                 statusImageView?.setStatus(false)
                 var stopOnError = true
-                if let stopOnErrorSwitch = self?.stopOnErrorSwitch? {
+                if let stopOnErrorSwitch = self?.stopOnErrorSwitch {
                     stopOnError = stopOnErrorSwitch.on
                 }
                 if stopOnError {
@@ -125,10 +101,6 @@ class CompoundNetworkCallDemoViewController: BaseDemoViewController {
                 }
             }
         )
-    }
-
-    @IBAction func delayStepperAction(stepper:UIStepper) {
-        updateUI()
     }
 
 }
