+++ SwiftPromises Demo/CompoundNetworkCallDemoViewController.swift
@@ -1,5 +1,5 @@
 //
-//  ChainedNetworkCallDemoViewController.swift
+//  CompoundNetworkCallDemoViewController.swift
 //  SwiftPromises
 //
 //  Created by Douglas Sjoquist on 3/1/15.
@@ -7,21 +7,29 @@
 //
 
 import UIKit
+import SwiftPromises
 
-class ChainedNetworkCallDemoViewController: BaseDemoViewController {
+class CompoundNetworkCallDemoViewController: BaseDemoViewController {
 
     @IBOutlet var url1TextField:UITextField?
-    @IBOutlet var url2TextField:UITextField?
+    @IBOutlet var url2aTextField:UITextField?
+    @IBOutlet var url2bTextField:UITextField?
     @IBOutlet var url3TextField:UITextField?
+    @IBOutlet var delay2aStepper:UIStepper?
+    @IBOutlet var delay2bStepper:UIStepper?
+    @IBOutlet var delay2aLabel:UILabel?
+    @IBOutlet var delay2bLabel:UILabel?
     @IBOutlet var url1StatusImageView:UIImageView?
-    @IBOutlet var url2StatusImageView:UIImageView?
+    @IBOutlet var url2aStatusImageView:UIImageView?
+    @IBOutlet var url2bStatusImageView:UIImageView?
     @IBOutlet var url3StatusImageView:UIImageView?
     @IBOutlet var finalStatusImageView:UIImageView?
     @IBOutlet var stopOnErrorSwitch:UISwitch?
 
     override func viewWillAppear(animated: Bool) {
         url1TextField!.text = "http://cnn.com"
-        url2TextField!.text = "http://apple.com"
+        url2aTextField!.text = "http://apple.com"
+        url2bTextField!.text = "http://dilbert.com"
         url3TextField!.text = "http://nytimes.com"
 
         super.viewWillAppear(animated)
@@ -29,7 +37,8 @@ class ChainedNetworkCallDemoViewController: BaseDemoViewController {
 
     override func clearStatus() {
         url1StatusImageView!.setStatus(nil)
-        url2StatusImageView!.setStatus(nil)
+        url2aStatusImageView!.setStatus(nil)
+        url2bStatusImageView!.setStatus(nil)
         url3StatusImageView!.setStatus(nil)
         finalStatusImageView!.setStatus(nil)
     }
@@ -38,6 +47,13 @@ class ChainedNetworkCallDemoViewController: BaseDemoViewController {
         return true
     }
 
+    override func updateUI() {
+        super.updateUI()
+
+        delay2aLabel!.text =  "Delay2a: \(delay2aStepper!.value)"
+        delay2bLabel!.text =  "Delay2b: \(delay2bStepper!.value)"
+    }
+    
     override func start() {
         clearStatus()
         clearLog()
@@ -46,7 +62,13 @@ class ChainedNetworkCallDemoViewController: BaseDemoViewController {
 
         loadURL1StepPromise().then(
             { [weak self] (value) -> AnyObject? in
-                return self?.loadURL2StepPromise()
+                self?.log("loading promise2a and promise2b")
+                var promises:[Promise] = []
+                if let strongSelf = self {
+                    promises.append(strongSelf.loadURL2aStepPromise())
+                    promises.append(strongSelf.loadURL2bStepPromise())
+                }
+                return Promise.all(promises)
         } ).then(
             { [weak self] (value) -> AnyObject? in
                 return self?.loadURL3StepPromise()
@@ -66,20 +88,24 @@ class ChainedNetworkCallDemoViewController: BaseDemoViewController {
     }
 
     func loadURL1StepPromise() -> Promise {
-        return loadURLStepPromise(url1TextField!.text, statusImageView:url1StatusImageView!)
+        return loadURLStepPromise(url1TextField!.text, statusImageView:url1StatusImageView!, delay:0.0)
     }
 
-    func loadURL2StepPromise() -> Promise {
-        return loadURLStepPromise(url2TextField!.text, statusImageView:url2StatusImageView!)
+    func loadURL2aStepPromise() -> Promise {
+        return loadURLStepPromise(url2aTextField!.text, statusImageView:url2aStatusImageView!, delay:delay2aStepper!.value)
+    }
+
+    func loadURL2bStepPromise() -> Promise {
+        return loadURLStepPromise(url2bTextField!.text, statusImageView:url2bStatusImageView!, delay:delay2bStepper!.value)
     }
 
     func loadURL3StepPromise() -> Promise {
-        return loadURLStepPromise(url3TextField!.text, statusImageView:url3StatusImageView!)
+        return loadURLStepPromise(url3TextField!.text, statusImageView:url3StatusImageView!, delay:0.0)
     }
 
-    func loadURLStepPromise(urlString:String?, statusImageView:UIImageView?) -> Promise {
+    func loadURLStepPromise(urlString:String?, statusImageView:UIImageView?, delay:NSTimeInterval) -> Promise {
         var url:NSURL? = (urlString == nil) ? nil : NSURL(string:urlString!)
-        return loadURLPromise(url).then(
+        return loadURLPromise(url, delay:delay).then(
             { [weak self] (value) -> AnyObject? in
                 statusImageView?.setStatus(true)
                 var data:NSData? = value as? NSData
@@ -88,7 +114,7 @@ class ChainedNetworkCallDemoViewController: BaseDemoViewController {
             }, reject: { [weak self] (error) -> AnyObject? in
                 statusImageView?.setStatus(false)
                 var stopOnError = true
-                if let stopOnErrorSwitch = self?.stopOnErrorSwitch? {
+                if let stopOnErrorSwitch = self?.stopOnErrorSwitch {
                     stopOnError = stopOnErrorSwitch.on
                 }
                 if stopOnError {
@@ -100,6 +126,10 @@ class ChainedNetworkCallDemoViewController: BaseDemoViewController {
                 }
             }
         )
+    }
+
+    @IBAction func delayStepperAction(stepper:UIStepper) {
+        updateUI()
     }
 
 }
