+++ SwiftPromises Demo/Images.xcassets/AppIcon.appiconset/Contents.json
