+++ SwiftPromises Demo/Images.xcassets/Knob Cancel.imageset/Contents.json
