+++ SwiftPromises Demo/Images.xcassets/Knob Cancel.imageset/Knob Cancel.png
