+++ SwiftPromises Demo/Images.xcassets/Knob Valid Green.imageset/Knob Valid Green.png
