+++ SwiftPromises Demo/ViewController.swift
