+++ SwiftPromises/SwiftPromises.h
@@ -1,0 +1,19 @@
+//
+//  SwiftPromises.h
+//  SwiftPromises
+//
+//  Created by Steven Riggins on 4/2/15.
+//  Copyright (c) 2015 Ivy Gulch LLC. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for SwiftPromises.
+FOUNDATION_EXPORT double SwiftPromisesVersionNumber;
+
+//! Project version string for SwiftPromises Framework.
+FOUNDATION_EXPORT const unsigned char SwiftPromisesVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <SwiftPromises/PublicHeader.h>
+
+
