+++ SwiftPromisesDemoTests/SwiftPromisesDemoTests.swift
@@ -1,0 +1,35 @@
+//
+//  SwiftPromisesDemoTests.swift
+//  SwiftPromises
+//
+//  Created by Steven Riggins on 4/2/15.
+//  Copyright (c) 2015 Ivy Gulch LLC. All rights reserved.
+//
+
+import XCTest
+
+class SwiftPromisesDemoTests: XCTestCase {
+
+    override func setUp() {
+        super.setUp()
+        // Put setup code here. This method is called before the invocation of each test method in the class.
+    }
+    
+    override func tearDown() {
+        // Put teardown code here. This method is called after the invocation of each test method in the class.
+        super.tearDown()
+    }
+
+    func testExample() {
+        // This is an example of a functional test case.
+        XCTAssert(true, "Pass")
+    }
+
+    func testPerformanceExample() {
+        // This is an example of a performance test case.
+        self.measureBlock() {
+            // Put the code you want to measure the time of here.
+        }
+    }
+
+}
